+++ CS419-FinalProject/SPIMIndexerViASCII.cs
@@ -21,8 +21,8 @@ namespace CS419_FinalProject
         {
             this.path = path;
             // Read the file of stopwords
-            //stopwords = new HashSet<string>();
-            stopwords = GetStopwords();
+            stopwords = new HashSet<string>();
+            //stopwords = GetStopwords();
         }
 
         // Read all stopwords from file
@@ -47,7 +47,7 @@ namespace CS419_FinalProject
             Console.WriteLine("Index Construction using Single-pass in-memory Indexing");
 
             // Initialize the dictionary
-            Dictionary<string, List<int>> dict = new Dictionary<string, List<int>>(20011);
+            Dictionary<string, List<int>> dict = new Dictionary<string, List<int>>(252143);
             // Count the size (in bytes) of the values (term, docId, frequency) added to dictionary and postings lists
             int count = 0;
             // Count the number of blocks written to disk
@@ -55,8 +55,11 @@ namespace CS419_FinalProject
 
             // Delete old files
             string[] filePaths = Directory.GetFiles("..//..//SPIMI//");
-            foreach (string filePath in filePaths)
+            foreach (string filePath in filePaths) {
+                if (filePath.Contains("index.txt") || filePath.Contains("index_map.txt") || filePath.Contains("index_length.txt"))
+                    continue;
                 File.Delete(filePath);
+            }
 
             // Count the total number of documents
             int docId = 0;
@@ -68,7 +71,7 @@ namespace CS419_FinalProject
                     string content = sr.ReadToEnd().ToLower();
                     if (content != null) {
                         content = Unicode2ASCII.Convert(content);
-                        SPIMI_Invert(content, docId, ref dict, ref count, "SPIMI_", ref blockId);
+                        SPIMI_Invert(content, docId, ref dict, ref count, "SPIMIvi_", ref blockId);
                     }
                     ++docId;
                 }
@@ -79,13 +82,13 @@ namespace CS419_FinalProject
                 Console.WriteLine("Block " + blockId.ToString() + ": Finished!");
                 List<string> sortedTerm = dict.Keys.ToList();
                 sortedTerm.Sort();
-                WriteBlockToDisk(sortedTerm, dict, "SPIMI_" + blockId.ToString());
+                WriteBlockToDisk(sortedTerm, dict, "SPIMIvi_" + blockId.ToString());
                 ++blockId;
             }
 
             // Merge all blocks
             Console.WriteLine("Merging blocks...");
-            MergeBlocks("SPIMI_", blockId, "..//..//SPIMI//index", docId);
+            MergeBlocks("SPIMIvi_", blockId, "..//..//SPIMI//indexvi", docId);
 
             Console.WriteLine("Inverted index successfully constructed.");
             return docId;
@@ -98,40 +101,30 @@ namespace CS419_FinalProject
             MatchCollection words = Tokenizer.TokenizeDoc(content, @"[a-zA-Z]+");
 
             // Add tokens to dictionary and postings lists
-            foreach (var word in words)
+
+            for (int i = 0; i < words.Count; ++i)
             {
-                string term = word.ToString().ToLower();
-                // Remove stopwords
-                if (stopwords.Contains(term))
-                    continue;
-                // Add term to dictionary
-                if (!dict.ContainsKey(term))
+
+                // Get the term
+                string term = words[i].ToString().ToLower();
+                // Add term to dictonary if it is not stopword
+                if (!stopwords.Contains(term))
+                    AddTermToDictionary(term, docId, 1, dict, ref count);
+
+                // Get the bigram
+                if (i + 1 < words.Count)
                 {
-                    dict.Add(term, new List<int>(200));
-                    count += term.Length << 1;
+
+                    term += " " + words[i + 1].ToString().ToLower();
+                    if (!stopwords.Contains(term))
+                        AddTermToDictionary(term, docId, 2, dict, ref count);
                 }
-                // Add posting directly to its postings list
-                bool found = false;
-                for (int i = 0; i < dict[term].Count; i += 2)
-                {
-                    // If docId already exist, increase the term frequency in doc
-                    if (dict[term][i] == docId)
-                    {
-                        ++dict[term][i + 1];
-                        found = true;
-                        break;
-                    }
-                }
-                // If docId does not exist, add it to postings list
-                if (!found)
-                {
-                    dict[term].Add(docId);
-                    dict[term].Add(1);
-                    count += 8;
-                }
+
             }
-            // We assume that only 1-2MB of memory is available
-            if (count >= 1048576)
+
+
+            // We assume that only 10MB of memory is available
+            if (count >= 10485760)
             {
                 Console.WriteLine("Block " + blockId.ToString() + ": Finished!");
                 // Sort terms
@@ -144,6 +137,37 @@ namespace CS419_FinalProject
                 count = 0;
                 ++blockId;
             }
+        }
+
+        // Add a term with a specific weight to the dictionary 
+        private int AddTermToDictionary(string term, int docId, int weight, Dictionary<string, List<int>> dict, ref int count)
+        {
+            // Add term to dictionary
+            if (!dict.ContainsKey(term))
+            {
+                dict.Add(term, new List<int>(200));
+                count += term.Length << 1;
+            }
+            // Add posting directly to its postings list
+            bool found = false;
+            for (int i = 0; i < dict[term].Count; i += 2)
+            {
+                // If docId already exist, increase the term frequency in doc
+                if (dict[term][i] == docId)
+                {
+                    dict[term][i + 1] += weight;
+                    found = true;
+                    break;
+                }
+            }
+            // If docId does not exist, add it to postings list
+            if (!found)
+            {
+                dict[term].Add(docId);
+                dict[term].Add(weight);
+                count += 8;
+            }
+            return count;
         }
 
         // Write the index of a block to disk in binary
